requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
tiktoken
InstructorEmbedding
sentence_transformers==2.2.2
audio-recorder-streamlit
SpeechRecognition
pydub
speechrecognition
pyaudio
pyttsx3
vosk
langchain_experimental
tabulate
googletrans==4.0.0-rc1
translate
deep-translator
streamlit-option-menu
